stat/pyproject.toml
[tool.poetry]
name = "stat"
version = "0.1.0"
description = ""
authors = ["DimaPekutko <dkefir03@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
pydantic = "^1.9.1"
python-dotenv=">=0.20.0"
boto3=">=1.24.30"
pyjwt=">=2.3.0"
pika = "^1.3.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
